requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
blinker==1.4
boto==2.42.0
celery==3.1.23
cffi==1.7.0
click==6.6
contextlib2==0.5.4
cryptography==1.4
cssmin==0.2.0
docutils==0.12
enum34==1.1.6
Fabric==1.10.2
Flask==0.11.1
Flask-Assets==0.11
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-S3==0.3.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
Flask-WTF==0.12
futures==3.0.5
hiredis==0.1.5
idna==2.1
ipaddress==1.0.16
itsdangerous==0.24
Jinja2==2.8
jmespath==0.9.0
jsmin==2.2.1
kombu==3.0.35
Mako==1.0.4
MarkupSafe==0.23
MySQL-python==1.2.5
paramiko==2.0.2
poster==0.8.1
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
PyJWT==0.4.3
pyOpenSSL==0.14
python-dateutil==2.5.3
python-editor==1.0.1
pytz==2016.6.1
raven==5.11.1
redis==2.10.3
s3transfer==0.1.1
six==1.10.0
sphinxalchemy==0.5.1
SQLAlchemy==0.9.8
uWSGI==2.0.13.1
webassets==0.11.1
Werkzeug==0.11.10
WTForms==2.1
Flask-Alembic==2.0.1
python-dateutil==2.6.0
gensim==0.12.4
korean
